requirements.txt
PySide6>=6.7
opencv-python-headless>=4.9
Pillow>=10.2
numpy>=1.26
pyinstaller>=6.6
